fixed tests

+++ tests/unit_tests/test_lm/test_mpiexec.py
@@ -171,33 +171,36 @@ class TestMPIExec(TestCase):
                     'mem'       : 0}
         ]
 
-        lm_mpiexec = MPIExec('', {}, None, None, None)
+        try:
+            lm_mpiexec = MPIExec('', {}, None, None, None)
 
-        host_file_expected = '%s/%s.hf' % (sandbox, uid)
-        self.assertFalse(os.path.isfile(host_file_expected))
+            host_file_expected = '%s/%s.hf' % (sandbox, uid)
+            self.assertFalse(os.path.isfile(host_file_expected))
 
-        host_file = lm_mpiexec._get_host_file(slots, uid, sandbox)
-        self.assertEqual(host_file_expected, host_file)
-        self.assertTrue(os.path.isfile(host_file))
+            host_file = lm_mpiexec._get_host_file(slots, uid, sandbox)
+            self.assertEqual(host_file_expected, host_file)
+            self.assertTrue(os.path.isfile(host_file))
 
-        # simple host file
-        with ru.ru_open(host_file) as hfd:
-            hfd_content = hfd.read()
-        self.assertEqual(hfd_content, 'node_A\nnode_B\n')
+            # simple host file
+            with ru.ru_open(host_file) as hfd:
+                hfd_content = hfd.read()
+            self.assertEqual(hfd_content, 'node_A\nnode_B\n')
 
-        # host file with "slots=" as delimiter for ranks
-        lm_mpiexec._get_host_file(slots, uid, sandbox, mode=1)
-        with ru.ru_open(host_file) as hfd:
-            hfd_content = hfd.read()
-        self.assertEqual(hfd_content, 'node_A slots=2\nnode_B slots=1\n')
+            # host file with "slots=" as delimiter for ranks
+            lm_mpiexec._get_host_file(slots, uid, sandbox, mode=1)
+            with ru.ru_open(host_file) as hfd:
+                hfd_content = hfd.read()
+            self.assertEqual(hfd_content, 'node_A slots=2\nnode_B slots=1\n')
 
-        # host file with ":" as delimiter for ranks
-        lm_mpiexec._get_host_file(slots, uid, sandbox, mode=2)
-        with ru.ru_open(host_file) as hfd:
-            hfd_content = hfd.read()
-        self.assertEqual(hfd_content, 'node_A:2\nnode_B:1\n')
+            # host file with ":" as delimiter for ranks
+            lm_mpiexec._get_host_file(slots, uid, sandbox, mode=2)
+            with ru.ru_open(host_file) as hfd:
+                hfd_content = hfd.read()
+            self.assertEqual(hfd_content, 'node_A:2\nnode_B:1\n')
 
-        os.unlink(host_file)
+        finally:
+            if os.path.isfile(host_file_expected):
+                os.unlink(host_file)
 
     # --------------------------------------------------------------------------
     #
@@ -300,7 +303,7 @@ class TestMPIExec(TestCase):
                 f_name   = test_case[3]
                 with ru.ru_open(f_name) as fd:
                     self.assertEqual(fd.readlines(), f_layout)
-                os.unlink(f_name)
+              # os.unlink(f_name)
 
     # --------------------------------------------------------------------------
     #
